Working on making the PQ from int to double

+++ PQ.h
@@ -4,7 +4,7 @@ typedef struct pqueue *PQ;
 PQ      PQinit(int maxN);
 void    PQfree(PQ pq);
 int     PQempty(PQ pq);
-void    PQinsert(PQ pq, int *mindist, int node);
-int     PQextractMin(PQ pq, int *mindist);
-void    PQchange (PQ pq, int *mindist, int k);
+void    PQinsert(PQ pq, double *mindist, int node);
+int     PQextractMin(PQ pq, double *mindist);
+void    PQchange (PQ pq, double *mindist, int k);
 #endif
